update bench tests to use simapp
